address review comments: move modeling blocks, inline sink-bias defaults
